+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatTreeUI.java
@@ -420,6 +420,9 @@ public class FlatTreeUI
 
 	/** @since 2 */
 	protected Object applyStyleProperty( String key, Object value ) {
+		if( "rowHeight".equals( key ) && value instanceof Integer )
+			value = UIScale.scale( (Integer) value );
+
 		return FlatStylingSupport.applyToAnnotatedObjectOrComponent( this, tree, key, value );
 	}